pyproject.toml
[tool.poetry]
name = "o1a-visa-analyzer"
version = "0.1.0"
description = "O-1A Visa eligibility analyzer"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
openai = "^1.0.0"
python-multipart = "^0.0.5"
tenacity = "^8.0.1"
PyPDF2 = "^1.26.0"
python-docx = "^0.8.11"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"